+++ tests/unit/server/services/ssh_key_cluster_1526/__init__.py
@@ -1,0 +1,1 @@
+"""Bug #1526: SSHKeyManager get_public_key / assign_key_to_host cluster-awareness."""